Make clang++ work

+++ inputcompare.cpp
@@ -3,14 +3,14 @@
 int inputcompare(void *type, char character)
 {
 #ifdef INITNCURSESNOW
-	if((int)type == character)
+	if((int)(size_t)type == character)
 	{
 		return 1;
 	}
 #endif
 
 #ifdef INITLIBTCODNOW
-	if((char)type == character)
+	if((char)(size_t)type == character)
 	{
 		return 1;
 	}
